feat: Port package and tests to Deno first
