+++ main/hooks/README.md
@@ -1,0 +1,37 @@
+# Hooks
+
+[![CI GitHub action](https://github.com/feathersjs/hooks/workflows/Node%20CI/badge.svg)](https://github.com/feathersjs/hooks/actions?query=workflow%3A%22Node+CI%22)
+[![Deno CI](https://github.com/feathersjs/hooks/actions/workflows/deno.yml/badge.svg)](https://github.com/feathersjs/hooks/actions/workflows/deno.yml)
+[![Download Status](https://img.shields.io/npm/dm/@feathersjs/hooks.svg?style=flat-square)](https://www.npmjs.com/package/@feathersjs/hooks)
+
+`@feathersjs/hooks` brings Koa style middleware to any asynchronous JavaScript and TypeScript function or class.
+
+## Installation
+
+### Node
+
+```
+npm install @feathersjs/hooks --save
+```
+
+```js
+const { hooks } = require('@feathersjs/hooks');
+
+import { hooks } from '@feathersjs/hooks';
+```
+
+### Deno
+
+```js
+import { hooks } from 'https://deno.land/x/hooks@x.x.x/deno/index.ts';
+```
+
+## Documentation
+
+See [feathersjs/hooks](https://github.com/feathersjs/hooks/) for the complete documentation.
+
+## License
+
+Copyright (c) 2020
+
+Licensed under the [MIT license](LICENSE).
