+++ main/hooks/test/dependencies.ts
@@ -1,0 +1,9 @@
+export * from 'https://deno.land/std@0.114.0/testing/asserts.ts';
+
+export const it = (name: string, fn: () => any, only: boolean = false) => Deno.test({
+  only,
+  name,
+  fn
+});
+
+it.only = (name: string, fn: () => any) => it(name, fn, true);
