+++ packages/hooks/test/index.test.ts
@@ -1,0 +1,7 @@
+import * as assert from 'assert';
+
+describe('no tests here', () => {
+  it('passes', () => {
+    assert.ok(true, 'Magic!');
+  });
+});
